artery/PyHaemo-main/pyproject.toml
[project]
name = "splico"
readme = "README.md"
authors = [
    { name = "Jochen Hinz", email = "jochen.hinz@epfl.ch" },
    { name = "Fabio Marcinno", email = "fabio.marcinno@epfl.ch" },
]
requires-python = '>=3.12'
dependencies = [
    "numpy >=1.26.4, <2", 
    "scipy >=1.12.0, <2",
    "nutils >=8.8, <9",
    "numba >=0.60.0, <1",
    "vtk >=9.3.1, <10",
    "pyvista >=0.44.1, <1",
    "pygmsh >=7.1.17, <8",
    "networkx >=3.4.2, <4"
]
dynamic = ["description", "version"]

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"
